Vagrant: Use private network by default.
Doing so improves the usability of the quickstart, especially for Windows users. Now, the ecosystem should start for all users, even when they do not have admin privileges or their network does not support bridging. It is also secure by default, because the setup wizard can only be run from their host.

For development we can use the vagrant.rb

This reflects feedback from users on Windows.